feat: update balances on demand
