+++ src/db/balances-groups.ts
@@ -97,6 +97,22 @@ export function toStorage(balancesGroups: BalancesGroup[]) {
   return balanceGroupsStorable
 }
 
+export async function selectAreBalancesStaleByFromAddress(client: PoolClient, fromAddress: string) {
+  const balancesRes = await client.query(
+    `
+    select
+      timestamp < now() - interval '5 minutes' as is_stale
+    from balances_groups
+    where from_address = $1
+    order by timestamp desc
+    limit 1;
+  `,
+    [fromAddress.toLowerCase()],
+  )
+
+  return balancesRes.rows[0].is_stale
+}
+
 export async function selectLastBalancesGroupsByFromAddress(client: PoolClient, fromAddress: string) {
   const balancesRes = await client.query(
     `
