+++ src/lib/balance.ts
@@ -1,19 +1,32 @@
+import { adapterById } from '@adapters'
+import { selectDefinedAdaptersContractsProps } from '@db/adapters'
+import type { Balance as BalanceStore } from '@db/balances'
+import { updateBalances as updateDBBalances } from '@db/balances'
+import { getAllContractsInteractions, groupContracts } from '@db/contracts'
 import type {
   Balance,
+  BalancesConfig,
   BalancesContext,
+  BalancesGroup,
   BaseBalance,
   BaseContract,
   ExcludeRawContract,
   GetContractsHandler,
   PricedBalance,
 } from '@lib/adapter'
+import { groupBy, groupBy2, keyBy2 } from '@lib/array'
 import type { Category } from '@lib/category'
+import type { Chain } from '@lib/chains'
 import { ADDRESS_ZERO } from '@lib/contract'
 import { getERC20BalanceOf } from '@lib/erc20'
+import { sum } from '@lib/math'
 import { multicall } from '@lib/multicall'
+import { getPricedBalances } from '@lib/price'
 import { providers } from '@lib/providers'
 import type { Token } from '@lib/token'
 import { isNotNullish } from '@lib/type'
+import type { PoolClient } from 'pg'
+import { v4 as uuidv4 } from 'uuid'
 
 export async function getBalances(ctx: BalancesContext, contracts: BaseContract[]) {
   const coins: Token[] = []
@@ -294,6 +307,8 @@ export function fmtBalanceBreakdown(balance: PricedBalance): PricedBalance & Bal
   }
 }
 
+export const BALANCE_UPDATE_THRESHOLD = 5 * 60 * 1000
+
 /**
  * At the moment, balances are considered "stale" if they haven't been updated in the last x minutes.
  * Later, we can use more advanced strategies using transactions events, scheduled updates etc
@@ -302,7 +317,182 @@ export function fmtBalanceBreakdown(balance: PricedBalance): PricedBalance & Bal
 export function areBalancesStale(lastUpdateTimestamp: number) {
   const now = new Date().getTime()
 
-  const updateInterval = 5 * 60 * 1000
+  return now - lastUpdateTimestamp > BALANCE_UPDATE_THRESHOLD
+}
 
-  return now - lastUpdateTimestamp > updateInterval
+type ExtendedBalance = (Balance | PricedBalance) & {
+  adapterId: string
+  groupIdx: number
+}
+
+interface BalancesGroupExtended {
+  balances: ExtendedBalance[]
+  healthFactor?: number
+}
+
+interface ExtendedBalancesConfig extends BalancesConfig {
+  adapterId: string
+  chain: Chain
+  groups: BalancesGroupExtended[]
+}
+
+/**
+ *
+ * @param client must be connected
+ * @param address
+ */
+export async function updateBalances(client: PoolClient, address: `0x${string}`) {
+  // Fetch all protocols (with their associated contracts) that the user interacted with
+  // and all unique tokens he received
+  const [contracts, adaptersContractsProps] = await Promise.all([
+    getAllContractsInteractions(client, address),
+    selectDefinedAdaptersContractsProps(client),
+  ])
+
+  const contractsByAdapterIdChain = groupBy2(contracts, 'adapterId', 'chain')
+  const adaptersContractsPropsByIdChain = keyBy2(adaptersContractsProps, 'id', 'chain')
+  // add adapters with contracts_props, even if there was no user interaction with any of the contracts
+  for (const adapter of adaptersContractsProps) {
+    if (!contractsByAdapterIdChain[adapter.id]) {
+      contractsByAdapterIdChain[adapter.id] = {}
+    }
+    if (!contractsByAdapterIdChain[adapter.id][adapter.chain]) {
+      contractsByAdapterIdChain[adapter.id][adapter.chain] = []
+    }
+  }
+
+  const adapterIds = Object.keys(contractsByAdapterIdChain)
+  // list of all [adapterId, chain]
+  const adapterIdsChains = adapterIds.flatMap((adapterId) =>
+    Object.keys(contractsByAdapterIdChain[adapterId]).map((chain) => [adapterId, chain] as [string, Chain]),
+  )
+
+  console.log('Interacted with protocols:', adapterIds)
+
+  // Run adapters `getBalances` only with the contracts the user interacted with
+  const adaptersBalancesConfigsRes = await Promise.all(
+    adapterIdsChains.map(async ([adapterId, chain]) => {
+      const adapter = adapterById[adapterId]
+      if (!adapter) {
+        console.error(`Could not find adapter with id`, adapterId)
+        return
+      }
+      const handler = adapter[chain]
+      if (!handler) {
+        console.error(`Could not find chain handler for`, [adapterId, chain])
+        return
+      }
+
+      try {
+        const hrstart = process.hrtime()
+
+        const contracts = groupContracts(contractsByAdapterIdChain[adapterId][chain]) || []
+        const props = adaptersContractsPropsByIdChain[adapterId]?.[chain]?.contractsProps || {}
+
+        const ctx: BalancesContext = { address, chain, adapterId }
+
+        const balancesConfig = await handler.getBalances(ctx, contracts, props)
+
+        const hrend = process.hrtime(hrstart)
+
+        const balancesLength = balancesConfig.groups.reduce((acc, group) => acc + (group.balances?.length || 0), 0)
+        console.log(
+          `[${adapterId}][${chain}] getBalances ${contractsByAdapterIdChain[adapterId][chain].length} contracts, found ${balancesLength} balances in %ds %dms`,
+          hrend[0],
+          hrend[1] / 1000000,
+        )
+
+        const extendedBalancesConfig: ExtendedBalancesConfig = {
+          ...balancesConfig,
+          // Tag balances with adapterId abd groupIdx
+          groups: balancesConfig.groups.map((balancesGroup, groupIdx) => ({
+            ...balancesGroup,
+            balances: balancesGroup.balances.map((balance) => ({ ...balance, adapterId, groupIdx })),
+          })),
+          adapterId,
+          chain,
+        }
+
+        return extendedBalancesConfig
+      } catch (error) {
+        console.error(`[${adapterId}][${chain}]: Failed to getBalances`, error)
+        return
+      }
+    }),
+  )
+
+  const adaptersBalancesConfigs = adaptersBalancesConfigsRes.filter(isNotNullish)
+
+  // Ungroup balances to make only 1 call to the price API
+  const balances: ExtendedBalance[] = []
+  for (const balancesConfig of adaptersBalancesConfigs) {
+    for (const group of balancesConfig.groups) {
+      for (const balance of group.balances) {
+        balances.push(balance)
+      }
+    }
+  }
+
+  const sanitizedBalances = sanitizeBalances(balances)
+
+  const hrstart = process.hrtime()
+
+  const pricedBalances = await getPricedBalances(sanitizedBalances)
+
+  const hrend = process.hrtime(hrstart)
+
+  console.log(
+    `getPricedBalances ${sanitizedBalances.length} balances, found ${pricedBalances.length} balances in %ds %dms`,
+    hrend[0],
+    hrend[1] / 1000000,
+  )
+
+  // Group balances back by adapter/chain
+  const pricedBalancesByAdapterIdChain = groupBy2(pricedBalances, 'adapterId', 'chain')
+
+  const now = new Date()
+
+  const balancesGroupsStore: BalancesGroup[] = []
+  const balancesStore: BalanceStore[] = []
+
+  for (const balanceConfig of adaptersBalancesConfigs) {
+    const pricedBalances = pricedBalancesByAdapterIdChain[balanceConfig.adapterId]?.[balanceConfig.chain]
+    if (!pricedBalances || pricedBalances.length === 0) {
+      continue
+    }
+
+    const balancesByGroupIdx = groupBy(pricedBalances.filter(isNotNullish), 'groupIdx')
+
+    for (let groupIdx = 0; groupIdx < balanceConfig.groups.length; groupIdx++) {
+      const balances = balancesByGroupIdx[groupIdx]
+      if (!balances || balances.length === 0) {
+        continue
+      }
+
+      const id = uuidv4()
+
+      const groupBalances = balances.map((balance) => ({ groupId: id, ...fmtBalanceBreakdown(balance) }))
+
+      for (const balance of groupBalances) {
+        balancesStore.push(balance)
+      }
+
+      const balancesGroup: BalancesGroup = {
+        id,
+        fromAddress: address,
+        adapterId: balanceConfig.adapterId,
+        chain: balanceConfig.chain,
+        balanceUSD: sum(groupBalances.map((balance) => balance.balanceUSD || 0)),
+        rewardUSD: sum(groupBalances.map((balance) => balance.rewardUSD || 0)),
+        debtUSD: sum(groupBalances.map((balance) => balance.debtUSD || 0)),
+        timestamp: now,
+        healthFactor: balanceConfig.groups[groupIdx].healthFactor,
+      }
+
+      balancesGroupsStore.push(balancesGroup)
+    }
+  }
+
+  // Update balances
+  return updateDBBalances(client, address, balancesGroupsStore, balancesStore)
 }
